+++ app/src/main/java/org/simple/clinic/registration/name/RegistrationFullNameScreen.kt
@@ -143,7 +143,7 @@ class RegistrationFullNameScreen(
   }
 
   override fun openRegistrationPinEntryScreen(registrationEntry: OngoingRegistrationEntry) {
-    router.push(RegistrationPinScreenKey(registrationEntry).wrap())
+    router.push(RegistrationPinScreenKey(registrationEntry))
   }
 
   interface Injector {