+++ app/src/main/java/org/simple/clinic/registration/pin/RegistrationPinScreenKey.kt
@@ -2,17 +2,16 @@ package org.simple.clinic.registration.pin
 
 import kotlinx.android.parcel.IgnoredOnParcel
 import kotlinx.android.parcel.Parcelize
-import org.simple.clinic.R
-import org.simple.clinic.router.screen.FullScreenKey
+import org.simple.clinic.navigation.v2.ScreenKey
 import org.simple.clinic.user.OngoingRegistrationEntry
 
 @Parcelize
 data class RegistrationPinScreenKey(
     val registrationEntry: OngoingRegistrationEntry
-) : FullScreenKey {
+) : ScreenKey() {
 
   @IgnoredOnParcel
   override val analyticsName = "Registration PIN Entry"
 
-  override fun layoutRes() = R.layout.screen_registration_pin
+  override fun instantiateFragment() = RegistrationPinScreen()
 }